Move Atmosphere feature from experimental to core library.
