+++ worldwind-examples/src/main/java/gov/nasa/worldwindx/BasicGlobeActivity.java
@@ -12,7 +12,7 @@ import gov.nasa.worldwind.WorldWindow;
 import gov.nasa.worldwind.globe.BasicElevationCoverage;
 import gov.nasa.worldwind.layer.BackgroundLayer;
 import gov.nasa.worldwind.layer.BlueMarbleLandsatLayer;
-import gov.nasa.worldwindx.experimental.AtmosphereLayer;
+import gov.nasa.worldwind.layer.atmosphere.AtmosphereLayer;
 
 /**
  * Creates a simple view of a globe with touch navigation and a few layers.
