+++ worldwind/src/main/java/gov/nasa/worldwind/layer/atmosphere/DrawableGroundAtmosphere.java
@@ -3,7 +3,7 @@
  * National Aeronautics and Space Administration. All Rights Reserved.
  */
 
-package gov.nasa.worldwindx.experimental;
+package gov.nasa.worldwind.layer.atmosphere;
 
 import android.opengl.GLES20;
