+++ worldwind/src/main/res/drawable/dnb_land_ocean_ice_2012.png
