+++ worldwind/src/main/res/raw/gov_nasa_worldwind_groundprogram_frag.glsl
