+++ worldwind/src/main/res/raw/gov_nasa_worldwind_skyprogram_frag.glsl
